Refactored all the names so that they are not index.js (#18)
