+++ src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import './App.css'
 import router from 'config/router'
 import { Provider } from 'react-redux'
-import store from 'store/index'
+import store from 'store/store'
 
 class App extends Component {
     render() {
