+++ src/config.json
@@ -1,0 +1,3 @@
+{
+  "GOOGLE_CLIENT_ID": "206945578523-0h8t8i7k5d09j0vg31ncspa4pbrddff6.apps.googleusercontent.com"
+}
