+++ src/config/router.js
@@ -9,9 +9,9 @@ import {
 import createHistory from 'history/createBrowserHistory'
 
 import Layout from 'layout'
-import GoogleLoginPage from 'views/google-login'
-import Arrange from 'views/arrange'
-import AddArrange from 'views/addarrange'
+import GoogleLoginPage from 'views/google-login/google-login'
+import Arrangement from 'views/arrangement/arrangement'
+import AddArrangement from 'views/addarrangement/addarrangement'
 
 const history = createHistory()
 
@@ -20,8 +20,8 @@ const router = (
     <Layout>
       <Switch>
         <Route exact path='/' component={GoogleLoginPage} />
-        <Route exact path='/arrangement' component={Arrange} />
-        <Route exact path='/addarrangement' component={AddArrange} />
+        <Route exact path='/arrangement' component={Arrangement} />
+        <Route exact path='/addarrangement' component={AddArrangement} />
       </Switch>
     </Layout>
   </Router>
