+++ src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
-import store from 'store/index'
+import store from 'store/store'
 
 function handleChange () {
   console.log(store.getState())
