+++ src/reducers/reducers.js
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux'
 
-import real from './real'
-import account from './account'
-import arrangements from './arrangements'
+import real from './real/real'
+import account from './account/account'
+import arrangements from './arrangements/arrangements'
 
 const rootReducer = combineReducers({
   real,
