+++ src/store/store.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
-import rootReducer from 'reducers'
+import rootReducer from 'reducers/reducers'
 
 const initialStore = {}
